feat(v1): introduce <V1App /> and the V1_REDESIGN boundary read in App.tsx
